forge fmt

+++ evm/src/interfaces/IWormhole.sol
@@ -25,10 +25,8 @@ interface IWormhole {
         uint64 sequence;
         uint8 consistencyLevel;
         bytes payload;
-
         uint32 guardianSetIndex;
         Signature[] signatures;
-
         bytes32 hash;
     }
 
@@ -36,7 +34,6 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
-
         address newContract;
     }
 
@@ -44,7 +41,6 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
-
         GuardianSet newGuardianSet;
         uint32 newGuardianSetIndex;
     }
@@ -53,7 +49,6 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
-
         uint256 messageFee;
     }
 
@@ -61,7 +56,6 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
-
         uint256 amount;
         bytes32 recipient;
     }
@@ -69,12 +63,13 @@ interface IWormhole {
     struct RecoverChainId {
         bytes32 module;
         uint8 action;
-
         uint256 evmChainId;
         uint16 newChainId;
     }
 
-    event LogMessagePublished(address indexed sender, uint64 sequence, uint32 nonce, bytes payload, uint8 consistencyLevel);
+    event LogMessagePublished(
+        address indexed sender, uint64 sequence, uint32 nonce, bytes payload, uint8 consistencyLevel
+    );
     event ContractUpgraded(address indexed oldContract, address indexed newContract);
     event GuardianSetAdded(uint32 indexed index);
 
@@ -86,15 +81,25 @@ interface IWormhole {
 
     function initialize() external;
 
-    function parseAndVerifyVM(bytes calldata encodedVM) external view returns (VM memory vm, bool valid, string memory reason);
+    function parseAndVerifyVM(bytes calldata encodedVM)
+        external
+        view
+        returns (VM memory vm, bool valid, string memory reason);
 
     function verifyVM(VM memory vm) external view returns (bool valid, string memory reason);
 
-    function verifySignatures(bytes32 hash, Signature[] memory signatures, GuardianSet memory guardianSet) external pure returns (bool valid, string memory reason);
+    function verifySignatures(
+        bytes32 hash,
+        Signature[] memory signatures,
+        GuardianSet memory guardianSet
+    ) external pure returns (bool valid, string memory reason);
 
     function parseVM(bytes memory encodedVM) external pure returns (VM memory vm);
 
-    function quorum(uint numGuardians) external pure returns (uint numSignaturesRequiredForQuorum);
+    function quorum(uint256 numGuardians)
+        external
+        pure
+        returns (uint256 numSignaturesRequiredForQuorum);
 
     function getGuardianSet(uint32 index) external view returns (GuardianSet memory);
 
@@ -120,15 +125,30 @@ interface IWormhole {
 
     function nextSequence(address emitter) external view returns (uint64);
 
-    function parseContractUpgrade(bytes memory encodedUpgrade) external pure returns (ContractUpgrade memory cu);
+    function parseContractUpgrade(bytes memory encodedUpgrade)
+        external
+        pure
+        returns (ContractUpgrade memory cu);
 
-    function parseGuardianSetUpgrade(bytes memory encodedUpgrade) external pure returns (GuardianSetUpgrade memory gsu);
+    function parseGuardianSetUpgrade(bytes memory encodedUpgrade)
+        external
+        pure
+        returns (GuardianSetUpgrade memory gsu);
 
-    function parseSetMessageFee(bytes memory encodedSetMessageFee) external pure returns (SetMessageFee memory smf);
+    function parseSetMessageFee(bytes memory encodedSetMessageFee)
+        external
+        pure
+        returns (SetMessageFee memory smf);
 
-    function parseTransferFees(bytes memory encodedTransferFees) external pure returns (TransferFees memory tf);
+    function parseTransferFees(bytes memory encodedTransferFees)
+        external
+        pure
+        returns (TransferFees memory tf);
 
-    function parseRecoverChainId(bytes memory encodedRecoverChainId) external pure returns (RecoverChainId memory rci);
+    function parseRecoverChainId(bytes memory encodedRecoverChainId)
+        external
+        pure
+        returns (RecoverChainId memory rci);
 
     function submitContractUpgrade(bytes memory _vm) external;
